+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 	## __WORK IN PROGRESS__
 -->
 
-## __WORK IN PROGRESS__
+## __WORK IN PROGRESS__ - Lucy
+* (@foxriver76) the UI upgrade now runs as the same user as the js-controller
 * (@foxriver76) fixed `iob validate` command and `setup custom` migration
 
 ## 7.0.1 (2024-10-21) - Lucy